Объединяет типы для api в файле server-data.ts
