+++ project/src/components/form-review/form-review.tsx
@@ -4,7 +4,7 @@ import { useAppDispatch, useAppSelector } from '../../hooks';
 import { postUserReviewAction } from '../../store/api-actions';
 import { clearErrorMessage, resetReviewSendSuccessStatus } from '../../store/app-data/app-data';
 import { getErrorMessage, getReviewSendStatus, getReviewSendSuccessStatus } from '../../store/app-data/selectors';
-import { ReviewData } from '../../types/app-data';
+import { ReviewData } from '../../types/server-data';
 import { CommentForm } from '../../types/reviews';
 import { toast } from 'react-toastify';
 import classNames from 'classnames';
