--- project/src/types/app-data.ts
@@ -1,7 +1,0 @@
-export type OfferId = string | undefined;
-
-export type ReviewData = {
-  id: number,
-  rating: number,
-  comment: string
-}
