--- project/src/types/favorite-data.ts
@@ -1,4 +1,0 @@
-export type FavoriteData = {
-  id: number,
-  status: 0 | 1
-}
