--- project/src/types/user-data.ts
@@ -1,10 +1,0 @@
-export type UserData = {
-  id: number;
-  email: string;
-  name: string;
-  avatarUrl: string;
-  isPro: boolean;
-  token: string;
-}
-
-export type UserEmail = string;
